cleanup unused constant files
